Extracted some code form long files
