+++ src/hooks/useClipboard.ts
@@ -1,9 +1,12 @@
 // src/hooks/useClipboard.ts
-import { useState } from 'react';
+import { useState, useEffect } from 'react';
 
 export function useClipboard() {
   const [copiedContent, setCopiedContent] = useState<string | null>(null);
 
+  /**
+   * Copies the given content to the clipboard and updates the state.
+   */
   const copyToClipboard = async (content: string) => {
     try {
       await navigator.clipboard.writeText(content);
@@ -13,6 +16,9 @@ export function useClipboard() {
     }
   };
 
+  /**
+   * Checks if the given content is currently in the clipboard.
+   */
   const isContentInClipboard = async (content: string) => {
     try {
       const clipboardText = await navigator.clipboard.readText();
@@ -22,6 +28,23 @@ export function useClipboard() {
       return false;
     }
   };
+
+  /**
+   * Automatically clears copied content if it's no longer in the clipboard.
+   * Runs an interval every 3 seconds to verify if the copied content is still present.
+   */
+  useEffect(() => {
+    if (!copiedContent) return;
+
+    const interval = setInterval(async () => {
+      const isStillInClipboard = await isContentInClipboard(copiedContent);
+      if (!isStillInClipboard) {
+        setCopiedContent(null);
+      }
+    }, 3000);
+
+    return () => clearInterval(interval); // Cleanup on unmount
+  }, [copiedContent]); // Runs whenever `copiedContent` changes
 
   return {
     copiedContent,
